+++ test/marinade-finance-instructions.spec.ts
@@ -2,6 +2,7 @@ import {
   MarinadeUtils,
   addLiquidity,
   claim,
+  createDirectedStakeVoteIx,
   deposit,
   depositStakeAccount,
   liquidUnstake,
@@ -139,9 +140,15 @@ describe('Marinade Finance', () => {
       const { transaction } = await deposit(
         marinadeProgram,
         TestWorld.SDK_USER.publicKey,
-        MarinadeUtils.solToLamports(0.01),
-        { directToValidatorVoteAddress: validatorVoteAddress }
+        MarinadeUtils.solToLamports(0.01)
       )
+
+      const voteIx = await createDirectedStakeVoteIx(
+        directedStakeSdk,
+        validatorVoteAddress
+      )
+      if (voteIx) transaction.instructions.push(voteIx)
+
       let transactionSignature: string
       try {
         transactionSignature = await TestWorld.PROVIDER.sendAndConfirm(
@@ -166,7 +173,7 @@ describe('Marinade Finance', () => {
         })
       )[0]
 
-      expect(voteRecord.account.validatorVote).toEqual(validatorVoteAddress)
+      expect(voteRecord.account.target).toEqual(validatorVoteAddress)
     })
 
     it('deposit SOL and redirect the stake', async () => {
@@ -190,9 +197,15 @@ describe('Marinade Finance', () => {
       const { transaction } = await deposit(
         marinadeProgram,
         TestWorld.SDK_USER.publicKey,
-        MarinadeUtils.solToLamports(0.01),
-        { directToValidatorVoteAddress: validatorVoteAddress2 }
+        MarinadeUtils.solToLamports(0.01)
       )
+
+      const voteIx = await createDirectedStakeVoteIx(
+        directedStakeSdk,
+        validatorVoteAddress2
+      )
+      if (voteIx) transaction.instructions.push(voteIx)
+
       const transactionSignature = await TestWorld.PROVIDER.sendAndConfirm(
         transaction,
         [],
@@ -211,7 +224,7 @@ describe('Marinade Finance', () => {
         })
       )[0]
 
-      expect(voteRecord?.account.validatorVote).toEqual(validatorVoteAddress2)
+      expect(voteRecord?.account.target).toEqual(validatorVoteAddress2)
     })
 
     it('deposit SOL and un-direct the stake', async () => {
@@ -235,6 +248,10 @@ describe('Marinade Finance', () => {
         TestWorld.SDK_USER.publicKey,
         MarinadeUtils.solToLamports(0.01)
       )
+
+      const voteIx = await createDirectedStakeVoteIx(directedStakeSdk)
+      if (voteIx) transaction.instructions.push(voteIx)
+
       const transactionSignature = await TestWorld.PROVIDER.sendAndConfirm(
         transaction,
         [],